Начинаем реализовывать фильтрацию изображений
